requirements.txt
python=3.12.9
uvicorn=0.34.0
fastapi=0.115.11
# streamlit=1.43.1
# pytorch=2.6
pytorch=2.5.1
cuda=12.4
accelerate=0.26.0
unstructured=0.16.25
modelscope=1.23.2
# chromadb=0.6.3
langchain
langserve[all]
langchain_community
langchain_huggingface
transformers
sentence-transformers
"unstructured[pdf]"
# error --> pip install opencv-python-headless
optimum
auto-gptq
gptqmodel
sse_starlette